completed exercise 4.23

+++ part4/blogList/backend/controllers/blogs.js
@@ -1,24 +1,54 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const jwt = require('jsonwebtoken')
+const User = require('../models/user')
+
+
 
  blogsRouter.get('/',async (request,response) => {
-   const blogs = await Blog.find({})
+  
+   const blogs = await Blog.find({}).populate('user',{username:1,name:1})
    response.json(blogs)
  })
 
- blogsRouter.post('/',(request,response,next) => {
+ blogsRouter.post('/',async (request,response,next) => {
     const body = request.body
-    const blog = new Blog(body)
-    blog.save().then(savedBlog => {
-        response.status(201).json(savedBlog)
+   const user = request.user
+   console.log('user:',user)
+   if(!user){
+          return response.status(400).json({error:'userId missing or not valid'})
+        }
+    const blog = new Blog( {
+      title:body.title,
+      author:body.author,
+      url:body.url,
+      likes:body.likes,
+      user:user._id
+
     })
-    .catch(error => next(error))
+    const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+    response.status(201).json(savedBlog)
+   
+    
  })
 
  blogsRouter.delete('/:id',async (request,response) => {
   const id = request.params.id
-  await Blog.findByIdAndDelete(id)
-  response.status(204).end()
+  const user = request.body
+  if(!user){
+          return response.status(400).json({error:'userId missing or not valid'})
+        }
+  const blog = await Blog.findById(id)
+  if(!blog){
+    return response.status(401).json({error:'invalid blog id'})
+  }
+  if(blog.user.toString() === user._id.toString()){
+    await Blog.findByIdAndDelete(id)
+    response.status(204).end()
+  }
+  
  })
 
  blogsRouter.put('/:id',async (request,response) => {
